Update Error Prone
This allows to re-enable the Mockito plugin.

There is a new check for potential int overflows before the value is
converted into a long, which actually found a bug (cf. previous commit).
For other cases where this check flags something the easiest way is to
just convert to long before the operation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40269 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMGBuiltins.java
@@ -217,7 +217,7 @@ public class SMGBuiltins {
             expressionEvaluator.writeValue(
                 currentState,
                 bufferMemory,
-                offset + (c * machineModel.getSizeofCharInBits()),
+                offset + ((long) c * machineModel.getSizeofCharInBits()),
                 CNumericTypes.SIGNED_CHAR,
                 ch,
                 cfaEdge);
